+++ CHANGELOG.md
@@ -1,8 +1,13 @@
-## 0.0.2
+## 0.0.4
 
-* First release.
+* New property `PointerEvent.isGlissandoEvent` that tells you if an event was synthesized by this
+  package. Only `PointerDownEvent` and `PointerCancelEvent` can be synthesized.
 
 ## 0.0.3
 
 * Provider more information to pointer cancel events.
 * New `enabled` parameter.
+
+## 0.0.2
+
+* First release.